requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-genai>=1.0.0
python-dotenv==1.0.1
pydantic==2.11.1
httpx==0.28.1
python-multipart==0.0.20
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
